+++ R/heatplot.itemfit.R
@@ -5,8 +5,17 @@
 #' @param object model object of class \code{itemfit}
 #' @param ... additional arguments
 #' @seealso \code{\link{GDINA}}, \code{\link{itemfit}}
-#' @export
 #'
+#' @examples
+#' \dontrun{
+#' dat <- sim10GDINA$simdat
+#' Q <- sim10GDINA$simQ
+#'
+#' fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
+#' ift <- itemfit(fit)
+#' heatplot(ift)
+#'}
+#' @export
 #'
 heatplot <- function(object,...){
   UseMethod("heatplot")