+++ NEWS.md
@@ -1,3 +1,12 @@
+# GDINA 1.9.9.9
+* This is a major update including a large number of new features. The `GDINA` function has been largely rewritten for both flexibility and speed. 
+  Users are now allowed to fit `MS-DINA` model, `Bugs` models, and define models by providing design matrix and link functions. For joint attribute
+  distribution, in addition to saturated and higher-order models, users are now allowed to fit independent model and loglinear model. Code for 
+  joint attribute distribution modelling has been restructured as well. Other major updates include model fit evaluation using M2 statistics and other
+  limited information measures, item-level model selection using likelihood ratio test and score test, classification accuracy evaluation indices, 
+  bootstrap standard error estimation, etc.
+* Note that due to the major updates, some results from this version may be slightly different from those using the previous versions. 
+
 # GDINA 1.4.2
 * Fixed     - a bug in `modelcomp` in the version 1.4.1
 * Changed   - adjusted p values are provided for DIF detection